requirements.txt
aioconsole==0.6.2
beautifulsoup4==4.12.2
bentoml==1.1.6
unstructured[all-docs]
fastapi
llama_index
faster_whisper
openai-whisper
firebase_admin==6.2.0
httpx==0.25.0
langchain
tiktoken
multion
pysqlite3-binary
numpy==1.26.0
openai==0.27.8
pgvector==0.2.3
pyaudio==0.2.13
pydantic==2.4.2
pydub==0.25.1
pytest==7.4.2
readerwriterlock==1.0.9
Requests==2.31.0
simpleaudio==1.0.4
SQLAlchemy==2.0.21
starlette==0.27
torch==2.1.0
torchaudio==2.1.0
python-dotenv==1.0.0
twilio==8.9.0
chromadb==0.4.13
SpeechRecognition==3.10.0
alembic==1.11.1
psycopg2-binary==2.9.9
pypdf==3.16.2
